Refactoring on Home compoennt and split functionality on separate components and utils functions
