+++ src/ui/Dropdown.jsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const Dropdown = styled.select`
+  padding: 8px;
+  border-radius: 6px;
+  border: 1px solid #ccc;
+  font-size: 1rem;
+`;
+
+export default Dropdown;
