+++ src/ui/FilterContainer.jsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const FilterContainer = styled.div`
+  margin-bottom: 20px;
+  display: flex;
+  gap: 10px;
+`;
+
+export default FilterContainer;
