+++ src/ui/Matchtitle.jsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+const MatchTitle = styled.p`
+  font-size: 1.2rem;
+  font-weight: 600;
+  color: #333;
+`;
+
+export default MatchTitle;
