+++ src/ui/TeamContainer.jsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const TeamContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  gap: 10px;
+`;
+
+export default TeamContainer;
